+++ migrations/1528395696_repo_name_index.down.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP INDEX IF EXISTS repo_name_idx;
+
+COMMIT;